# APItest/SwaggerUI api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SK API",
    "description": "SK智慧校园API文档",
    "license": {
      "name": "Apache 2.0",
      "url": "http://springdoc.org"
    },
    "version": "v114.514"
  },
...
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/sms/teacherController/saveOrUpdateTeacher": {
      "post": {
        "tags": [
          "Teacher"
        ],
        "summary": "保存或修改教师",
        "description": "保存或者修改教师信息",
        "operationId": "saveOrUpdateTeacher",
        "parameters": [
          {
            "name": "teacher",
            "in": "query",
            "description": "要保存或修改的教师JSON",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Teacher"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/teacherController/getTeachers/{pageNo}/{pageSize}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "分页带条件查询教师信息",
        "description": "分页带条件查询教师信息",
        "operationId": "getTeachers",
        "parameters": [
          {
            "name": "pageNo",
            "in": "path",
            "description": "页码数",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "path",
            "description": "页大小",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "teacher",
            "in": "query",
            "description": "查询的条件",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Teacher"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/sms/teacherController/deleteTeacher": {
      "delete": {
        "tags": [
          "Teacher"
        ],
        "summary": "删除教师",
        "description": "删除单个或者多个教师信息",
        "operationId": "deleteTeacher",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "要删除的教师编号的JSON数组",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultObject"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResultObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "消息",
            "example": "操作成功"
          },
          "data": {
            "type": "object",
            "description": "数据",
            "example": null
          },
          "ok": {
            "type": "boolean"
          }
        }
      },
      "Teacher": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "教师Id",
            "format": "int32",
            "example": 1
          },
          "tno": {
            "type": "string",
            "description": "教师工号",
            "example": "1001"
          },
          "name": {
            "type": "string",
            "description": "教师姓名",
            "example": "张三"
          },
          "gender": {
            "type": "string",
            "description": "教师性别",
            "example": "男"
          },
          "password": {
            "type": "string",
            "description": "教师密码",
            "example": "123456"
          },
          "email": {
            "type": "string",
            "description": "教师邮箱",
            "example": "2421361985"
          },
          "telephone": {
            "type": "string",
            "description": "教师电话",
            "example": "18888888888"
          },
          "address": {
            "type": "string",
            "description": "教师地址",
            "example": "广东省广州市"
          },
          "portrait": {
            "type": "string",
            "description": "教师头像",
            "example": "http://localhost:8080/sms/teacherController/getPortrait/1"
          },
          "portraitPath": {
            "type": "string",
            "description": "教师头像路径",
            "example": "http://localhost:8080/sms/teacherController/getPortrait/1"
          }
        }
      },
...
    }
  }
}